refactor: move manual utility tasks as flask cli module
